+++ packages/core/src/languages/html/attributes.ts
@@ -14,11 +14,11 @@ const GROUP_MAIN = "main";
 function newElementAttributeExpressions(): MangleExpression[] {
   return QUOTES_ARRAY.map((quote) => new NestedGroupExpression(
     `
-      (?<=\\<\\s*[a-zA-Z0-9]+\\s+)
+      (?<=\\<\\s*[a-zA-Z0-9]+\\s+[^>]*)
       (?<${GROUP_MAIN}>
         (?:
-          [^>${quote}]+
-          ${quote}[^${quote}]*${quote}
+          [^>]+
+          =${quote}[^${quote}]*${quote}
           \\s
         )*
         %s
@@ -27,12 +27,12 @@ function newElementAttributeExpressions(): MangleExpression[] {
           [^>]*
         )?
       )
-      (?=\\>)
+      (?=(\\/\\s*)?\\>)
     `,
     `
       (?<=\\s|^)
       (?<${GROUP_MAIN}>%s)
-      (?=\\=|\\s|$)
+      (?=\\=|\\s|\\/|\\>|$)
     `,
     GROUP_MAIN,
   ));